test: migrate SmokeStake Detox specs to Appium
Move stake and lending smoke tests to tests/smoke-appium/stake/, cut over CI from Detox to Appium SmokeStake jobs, and fix Appium page-object locators for stake/lending flows.
