+++ .github/workflows/run-e2e-smoke-tests-android.yml
@@ -24,23 +24,6 @@ permissions:
   id-token: write
 
 jobs:
-  stake-android-smoke:
-    if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokeStake')
-    strategy:
-      matrix:
-        split: [1]
-      fail-fast: false
-    uses: ./.github/workflows/run-e2e-workflow.yml
-    with:
-      test-suite-name: stake-android-smoke-${{ matrix.split }}
-      platform: android
-      test_suite_tag: 'SmokeStake'
-      split_number: ${{ matrix.split }}
-      total_splits: 1
-      changed_files: ${{ inputs.changed_files }}
-      runner_provider: ${{ inputs.runner_provider }}
-    secrets: inherit
-
   perps-android-smoke:
     if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokePerps')
     strategy:
@@ -181,7 +164,7 @@ jobs:
     name: Report Android Smoke Tests
     runs-on: ${{ inputs.runner_provider == 'namespace' && 'namespace-profile-metamask-ci-linux' || 'ubuntu-latest' }}
     # Appium migration: skip rollup when every Detox shard is skipped (e.g. AI selects
-    # an Appium-only tag like SmokeAccounts / SmokeSwap). Non-empty tags still trigger this workflow.
+    # an Appium-only tag like SmokeAccounts / SmokeSwap / SmokeStake). Non-empty tags still trigger this workflow.
     if: >-
       ${{
         !cancelled() &&
@@ -191,7 +174,6 @@ jobs:
         )
       }}
     needs:
-      - stake-android-smoke
       - perps-android-smoke
       - wallet-platform-android-smoke
       - network-abstraction-android-smoke
