+++ tests/page-objects/wallet/HomeSections.ts
@@ -3,7 +3,11 @@ import {
   Gestures,
   Matchers,
   EncapsulatedElementType,
+  encapsulated,
+  encapsulatedAction,
+  PlatformDetector,
 } from '../../framework';
+import PlaywrightMatchers from '../../framework/PlaywrightMatchers';
 import {
   WalletAssetSelectorsIDs,
   WalletAssetSelectorsRegex,
@@ -30,17 +34,30 @@ class TokensFullView {
   }
 
   get stakedEthereumAssetRow(): EncapsulatedElementType {
-    return Promise.resolve(
-      element(
-        by
-          .id(WalletAssetSelectorsIDs.STAKED_ETHEREUM)
-          .withDescendant(by.text(WalletViewSelectorsText.STAKED_ETHEREUM))
-          .withDescendant(
-            by.text(WalletAssetSelectorsText.STAKED_ETHEREUM_AMOUNT),
-          )
-          .withDescendant(by.text(WalletAssetSelectorsRegex.FIAT_BALANCE)),
-      ),
-    );
+    const assetId = WalletAssetSelectorsIDs.STAKED_ETHEREUM;
+    const stakedLabel = WalletViewSelectorsText.STAKED_ETHEREUM;
+    const amountText = WalletAssetSelectorsText.STAKED_ETHEREUM_AMOUNT;
+
+    return encapsulated({
+      detox: () =>
+        element(
+          by
+            .id(assetId)
+            .withDescendant(by.text(stakedLabel))
+            .withDescendant(by.text(amountText))
+            .withDescendant(by.text(WalletAssetSelectorsRegex.FIAT_BALANCE)),
+        ),
+      appium: () => {
+        if (PlatformDetector.isIOS()) {
+          return PlaywrightMatchers.getElementByIOSPredicate(
+            `name == '${assetId}' AND (label CONTAINS '${stakedLabel}' OR value CONTAINS '${stakedLabel}') AND (label CONTAINS '${amountText}' OR value CONTAINS '${amountText}')`,
+          );
+        }
+        return PlaywrightMatchers.getElementByXPath(
+          `//*[contains(@resource-id,'${assetId}')][descendant::*[@text='${stakedLabel}'] and descendant::*[@text='${amountText}']]`,
+        );
+      },
+    });
   }
 
   /**
@@ -63,8 +80,58 @@ class TokensFullView {
   }
 
   async expectStakedEthereumRowWithBalancesVisible(): Promise<void> {
-    await Assertions.expectElementToBeVisible(this.stakedEthereumAssetRow, {
-      description: 'Staked Ethereum row should display token and fiat balances',
+    await encapsulatedAction({
+      detox: async () => {
+        await Assertions.expectElementToBeVisible(this.stakedEthereumAssetRow, {
+          description:
+            'Staked Ethereum row should display token and fiat balances',
+          timeout: 60000,
+        });
+      },
+      appium: async () => {
+        if (PlatformDetector.isIOS()) {
+          await Assertions.expectElementToBeVisible(
+            encapsulated({
+              appium: () =>
+                PlaywrightMatchers.getElementByIOSPredicate(
+                  `label CONTAINS '${WalletViewSelectorsText.STAKED_ETHEREUM}' OR value CONTAINS '${WalletViewSelectorsText.STAKED_ETHEREUM}'`,
+                ),
+            }),
+            {
+              description: 'Staked Ethereum label should be visible',
+              timeout: 60000,
+            },
+          );
+          await Assertions.expectElementToBeVisible(
+            encapsulated({
+              appium: () =>
+                PlaywrightMatchers.getElementByIOSPredicate(
+                  `name == '${WalletAssetSelectorsIDs.STAKED_ETHEREUM}' AND (label CONTAINS '${WalletAssetSelectorsText.STAKED_ETHEREUM_AMOUNT}' OR value CONTAINS '${WalletAssetSelectorsText.STAKED_ETHEREUM_AMOUNT}')`,
+                ),
+            }),
+            {
+              description: 'Staked Ethereum amount (1 ETH) should be visible',
+              timeout: 30000,
+            },
+          );
+          return;
+        }
+
+        await Assertions.expectTextDisplayed(
+          WalletViewSelectorsText.STAKED_ETHEREUM,
+          {
+            timeout: 60000,
+            description: 'Staked Ethereum label should be visible',
+          },
+        );
+        await Assertions.expectTextDisplayed(
+          WalletAssetSelectorsText.STAKED_ETHEREUM_AMOUNT,
+          {
+            timeout: 30000,
+            description: 'Staked Ethereum amount (1 ETH) should be visible',
+          },
+        );
+      },
     });
   }
 }
