+++ tests/page-objects/wallet/TokensView.ts
@@ -5,7 +5,14 @@ import Gestures from '../../framework/Gestures';
 import Matchers from '../../framework/Matchers';
 import Utilities from '../../framework/Utilities';
 import NetworkManager from './NetworkManager';
-import { EncapsulatedElementType } from '../../framework';
+import {
+  encapsulated,
+  EncapsulatedElementType,
+  PlatformDetector,
+  encapsulatedAction,
+} from '../../framework';
+import PlaywrightMatchers from '../../framework/PlaywrightMatchers';
+import Assertions from '../../framework/Assertions';
 
 class TokensView {
   get networkFilter(): EncapsulatedElementType {
@@ -13,10 +20,24 @@ class TokensView {
   }
 
   earnCtaForToken(tokenSymbol: string): EncapsulatedElementType {
-    return Matchers.getElementIDWithAncestor(
-      SECONDARY_BALANCE_BUTTON_TEST_ID,
-      getAssetTestId(tokenSymbol),
-    );
+    const assetTestId = getAssetTestId(tokenSymbol);
+    return encapsulated({
+      detox: () =>
+        Matchers.getElementIDWithAncestor(
+          SECONDARY_BALANCE_BUTTON_TEST_ID,
+          assetTestId,
+        ),
+      appium: {
+        ios: () =>
+          PlaywrightMatchers.getElementByXPath(
+            `//*[@name='${assetTestId}']/descendant::*[@name='${SECONDARY_BALANCE_BUTTON_TEST_ID}']`,
+          ),
+        android: () =>
+          PlaywrightMatchers.getElementByXPath(
+            `//*[@resource-id='${assetTestId}' or contains(@resource-id,'${assetTestId}')]/descendant::*[@resource-id='${SECONDARY_BALANCE_BUTTON_TEST_ID}' or contains(@resource-id,'${SECONDARY_BALANCE_BUTTON_TEST_ID}')]`,
+          ),
+      },
+    });
   }
 
   async tapNetworkFilter(): Promise<void> {
@@ -46,10 +67,39 @@ class TokensView {
     timeout = 30000,
   ): Promise<void> {
     const assetTestId = getAssetTestId(tokenSymbol);
-    const zeroBalance = element(
-      by.text(`0 ${tokenSymbol}`).withAncestor(by.id(assetTestId)),
-    );
-    await waitFor(zeroBalance).not.toBeVisible().withTimeout(timeout);
+    await encapsulatedAction({
+      detox: async () => {
+        const zeroBalance = element(
+          by.text(`0 ${tokenSymbol}`).withAncestor(by.id(assetTestId)),
+        );
+        await waitFor(zeroBalance).not.toBeVisible().withTimeout(timeout);
+      },
+      appium: async () => {
+        await Assertions.expectElementToBeVisible(
+          Matchers.getElementByID(assetTestId),
+          {
+            timeout,
+            description: `${tokenSymbol} token row`,
+          },
+        );
+
+        const zeroBalanceText = `0 ${tokenSymbol}`;
+        const zeroBalanceXPath = PlatformDetector.isIOS()
+          ? `//*[@name='${assetTestId}']/descendant::*[@label='${zeroBalanceText}' or @name='${zeroBalanceText}' or @value='${zeroBalanceText}']`
+          : `//*[@resource-id='${assetTestId}' or contains(@resource-id,'${assetTestId}')]/descendant::*[@text='${zeroBalanceText}']`;
+
+        await Assertions.expectElementToNotBeVisible(
+          encapsulated({
+            appium: () =>
+              PlaywrightMatchers.getElementByXPath(zeroBalanceXPath),
+          }),
+          {
+            timeout,
+            description: `${tokenSymbol} balance should be non-zero`,
+          },
+        );
+      },
+    });
   }
 
   async tapToken(tokenSymbol: string): Promise<void> {
